requirements.txt
akshare>=1.10.0
pandas>=1.3.0
requests~=2.32.3
pyyaml~=6.0.2
duckdb>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
backtrader>=1.9.78.123
matplotlib>=3.5.0
